requirements.txt
python==3.11.1
matplotlib==3.7.1
numpy==1.23.5
Pillow==9.2.0
Pillow==10.0.1
streamlit==1.26.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
keras==2.12
pandas==2.0.0
sklearn==1.2.2
seaborn==0.12.2

